% Generated by roxygen2 (4.0.2): do not edit by hand
\name{noisy_cor}
\alias{noisy_cor}
\title{Generate noisy bivariate normal data and measure correlation}
\usage{
noisy_cor(num_genes, true_cor, noise_var, to_file = FALSE)
}
\arguments{
\item{num_genes}{Number of observations.}

\item{true_cor}{True correlation.}

\item{noise_var}{Variance of additional noise, independent for the
two variables.}

\item{to_file}{Whether to write the results to an RData file.}
}
\value{
Observerd correlation.
}
\description{
Generate noisy bivariate normal data and measure correlation
}

